Extract out synchronizer
