+++ packages/server/package.json
@@ -13,7 +13,9 @@
   ],
   "scripts": {
     "clean": "rimraf dist",
-    "predev": "wait-on ../core/dist/packages/core/src/index.d.ts",
+    "wait:synchronizer": "wait-on ../synchronizer/dist/packages/synchronizer/src/index.d.ts",
+    "wait:core": "wait-on ../core/dist/packages/core/src/index.d.ts",
+    "predev": "yarn wait:core && yarn wait:synchronizer",
     "dev": "tsdx watch --verbose",
     "build": "tsdx build",
     "test": "tsdx test",
@@ -33,6 +35,7 @@
   "types": "dist/packages/server/src/index.d.ts",
   "dependencies": {
     "@blitzjs/display": "0.12.0",
+    "@blitzjs/synchronizer": "0.12.0",
     "cross-spawn": "7.0.2",
     "detect-port": "1.3.0",
     "fast-glob": "3.2.2",
