+++ packages/server/src/index.ts
@@ -2,8 +2,7 @@ export {withBlitz} from './with-blitz'
 export {build} from './build'
 export {dev} from './dev'
 export {prod} from './prod'
-export {Manifest} from './synchronizer/pipeline/rules/manifest'
-export {ManifestLoader} from './synchronizer/manifest-loader'
 export {enhance} from './config'
 export {resolveBinAsync} from './resolve-bin-async'
+export {ManifestLoader} from './manifest-loader'
 export * from './rpc'
