+++ packages/synchronizer/README.md
@@ -1,0 +1,3 @@
+# `display`
+
+This package houses all files related to Blitz commandline output.
