+++ packages/synchronizer/src/diagram-error-and-view.png
