+++ packages/synchronizer/src/diagram-file-transform.png
