+++ packages/synchronizer/src/display.ts
