+++ packages/synchronizer/src/errors.ts
