+++ packages/synchronizer/src/index.ts
@@ -1,0 +1,2 @@
+export {synchronizeFiles} from './synchronizer'
+export {Manifest} from './pipeline/rules/manifest'
