+++ packages/synchronizer/src/pipeline/helpers/agnostic-source.ts
