+++ packages/synchronizer/src/pipeline/helpers/enrich-files.ts
