+++ packages/synchronizer/src/pipeline/helpers/file-cache.ts
