+++ packages/synchronizer/src/pipeline/helpers/idle-handler.test.ts
