+++ packages/synchronizer/src/pipeline/helpers/idle-handler.ts
