+++ packages/synchronizer/src/pipeline/helpers/work-optimizer.ts
