+++ packages/synchronizer/src/pipeline/rules/config/index.ts
