+++ packages/synchronizer/src/pipeline/rules/manifest/index.ts
