+++ packages/synchronizer/src/pipeline/rules/manifest/manifest.test.ts
