+++ packages/synchronizer/src/pipeline/rules/pages/index.ts
