+++ packages/synchronizer/src/pipeline/rules/relative/index.ts
