+++ packages/synchronizer/src/pipeline/rules/relative/relative-paths.test.ts
