+++ packages/synchronizer/src/pipeline/rules/write/index.ts
