+++ packages/synchronizer/src/vinyl-file.d.ts
